+++ client/src/pages/Admin/AdminCreateService.js
@@ -1,7 +1,7 @@
 import { Container, Flex } from '@chakra-ui/react';
 import React from 'react';
 import SideBar from '../../components/AdminPage/SideBar';
-import CreateServiceList from '../../components/AdminServiceList/CreateServiceCategory';
+import CreateServiceList from '../../components/AdminServiceList/CreateServiceList';
 
 const AdminCreateService = () => {
     return (
@@ -9,7 +9,6 @@ const AdminCreateService = () => {
             <Flex w='100vw'>
                 <SideBar />
                 <CreateServiceList />
-
             </Flex>
         </Container >
